chore(react-tree-grid): navigation up/down moves to next/previous row
